+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Add a link to the GitHub diff like
 ### Added
 - support for unpacking `*.tar.*` archives with two steps
 - app property `SetupTestFile` to explicitly specify the file
+- config property `DownloadProgress` to control progress info during app resource downloads
   which is used to determine if an app is installed or not
 - App: Dia
 - App: Gimp